#69 [REFACTOR] 공통 스타일 분리 및 페이지네이션 적용
