+++ src/pages/FriendPage/FriendReceivedList/FriendReceivedList.jsx
@@ -3,28 +3,31 @@ import { useQuery } from '@tanstack/react-query';
 import { getFriendList } from '../../../apis/friend.js';
 
 import useAuth from '../../../hooks/useAuth.jsx';
+import usePagination from '../../../hooks/usePagination.jsx';
 
 import { Avatar } from '../../../components/Avatar';
 import { Button } from '../../../components/Button';
 
-import * as S from './FriendReceivedList.style.jsx';
+import * as S from '../Common.Style.jsx';
 
 export default function FriendReceivedList() {
-  const { nickname } = useAuth();
-  const { data } = useQuery({
+  const { nickname: myNickname } = useAuth();
+  const { data, fetchNextPage } = usePagination({
     queryKey: ['friendList'],
-    queryFn: () => getFriendList(),
-    staleTime: 1000 * 60 * 5,
+    queryFn: ({ pageParam }) => getFriendList({ page: pageParam }),
   });
 
   if (!data) return null;
 
-  const friendList = data.filter(({ status }) => status === 'RECEIVED');
+  const friendList =
+    data && !data.pages.includes(undefined)
+      ? data.pages.flatMap((page) => page)
+      : [];
 
   return (
     <S.FriendLayout>
       <S.FriendContainer>
-        <S.Title>{nickname}님이 받은 친구 요청</S.Title>
+        <S.Title>{myNickname}님이 받은 친구 요청</S.Title>
         <S.FriendList>
           {friendList.map(({ id, nickname, profile_url: profileUrl }) => (
             <S.FriendItem
@@ -51,6 +54,21 @@ export default function FriendReceivedList() {
             </S.FriendItem>
           ))}
         </S.FriendList>
+        {friendList.length > 0 ? (
+          <Button
+            onClick={() => {
+              fetchNextPage();
+            }}
+            padding='10px'
+            borderRadius='5px'
+            buttonStyle='light'
+            hasBorder
+          >
+            더보기
+          </Button>
+        ) : (
+          <S.Text>받은 친구 요청이 없습니다</S.Text>
+        )}
       </S.FriendContainer>
     </S.FriendLayout>
   );
